Fixed a bug that prevented a few test cases from running.
There seems to be a bug in JUnitParams that allows it to silently ignore some
test cases when there is a conversion error for an argument instead of failing
with an error.

The coverage is now 100% for instructions and 97% for branches, but the one
branch that appears missing is really one Java semantics guarantee is
impossible to take! (this is probably a jacoco bug)

Reviewed at https://rbcommons.com/s/VMTurbo/r/6155/

git-svn-id: https://vcs.turbonomic.com/svn/analysis/trunk@32 3bb2320e-05c5-11dd-9ba1-b9e69423245e

+++ com.vmturbo.platform.analysis/src/test/java/com/vmturbo/platform/analysis/economy/CommoditySpecificationTest.java
@@ -164,8 +164,8 @@ public class CommoditySpecificationTest {
             output[c++] = new Object[]{invalidTypes[i],1,10};
         }
         for(int i = 0 ; i < invalidBounds.length ; ++i) {
-            output[c++] = new Object[]{0,invalidBounds[i],10};
-            output[c++] = new Object[]{0,1,invalidBounds[i]};
+            output[c++] = new Object[]{(short)0,invalidBounds[i],10};
+            output[c++] = new Object[]{(short)0,1,invalidBounds[i]};
         }
 
         return output;
